requirements.txt
# Local development dependencies (not Modal)
numpy<2
pillow
scipy
matplotlib
modal
